01_models fixed migrations. Fixed fixtures. Change php version in composer
